Restructure SAT controller actions
The SAT registration and dedicated Recovery token actions have been
decoupled into separate controllers. The shared logic is placed in a
trait. For now the constructors of both controllers are the same, and
they now live in the trait.

All controller actions have been given a developer friendly description
with a high level description of what the action is responsible for.
