+++ src/Surfnet/StepupSelfService/SelfServiceBundle/Controller/SelfAssertedTokensController.php
@@ -19,39 +19,32 @@
 namespace Surfnet\StepupSelfService\SelfServiceBundle\Controller;
 
 use Psr\Log\LoggerInterface;
+use Surfnet\StepupBundle\Service\LoaResolutionService;
+use Surfnet\StepupBundle\Service\SmsRecoveryTokenService;
 use Surfnet\StepupBundle\Value\PhoneNumber\InternationalPhoneNumber;
-use Surfnet\StepupMiddlewareClientBundle\Exception\NotFoundException;
-use Surfnet\StepupMiddlewareClientBundle\Identity\Dto\Identity;
 use Surfnet\StepupSelfService\SelfServiceBundle\Command\PromiseSafeStorePossessionCommand;
-use Surfnet\StepupSelfService\SelfServiceBundle\Command\RevokeRecoveryTokenCommand;
 use Surfnet\StepupSelfService\SelfServiceBundle\Command\SafeStoreAuthenticationCommand;
 use Surfnet\StepupSelfService\SelfServiceBundle\Command\SelfAssertedTokenRegistrationCommand;
 use Surfnet\StepupSelfService\SelfServiceBundle\Command\SendRecoveryTokenSmsAuthenticationChallengeCommand;
-use Surfnet\StepupSelfService\SelfServiceBundle\Command\SendRecoveryTokenSmsChallengeCommand;
 use Surfnet\StepupSelfService\SelfServiceBundle\Command\VerifySmsRecoveryTokenChallengeCommand;
-use Surfnet\StepupSelfService\SelfServiceBundle\Exception\LogicException;
 use Surfnet\StepupSelfService\SelfServiceBundle\Form\Type\AuthenticateSafeStoreType;
 use Surfnet\StepupSelfService\SelfServiceBundle\Form\Type\PromiseSafeStorePossessionType;
-use Surfnet\StepupSelfService\SelfServiceBundle\Form\Type\SendSmsChallengeType;
 use Surfnet\StepupSelfService\SelfServiceBundle\Form\Type\VerifySmsChallengeType;
 use Surfnet\StepupSelfService\SelfServiceBundle\Service\SecondFactorService;
+use Surfnet\StepupSelfService\SelfServiceBundle\Service\SelfAssertedTokens\AuthenticationRequestFactory;
 use Surfnet\StepupSelfService\SelfServiceBundle\Service\SelfAssertedTokens\RecoveryTokenService;
 use Surfnet\StepupSelfService\SelfServiceBundle\Service\SelfAssertedTokens\SafeStoreService;
-use Surfnet\StepupSelfService\SelfServiceBundle\Service\SmsRecoveryTokenService;
-use Surfnet\StepupSelfService\SelfServiceBundle\Service\SmsSecondFactorService;
-use Surfnet\StepupSelfService\SelfServiceBundle\Service\SmsSecondFactorServiceInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
 /**
- * @SuppressWarnings(PHPMD.TooManyPublicMethods) - Could be resolved by moving the non registration actions to a
- * controller of it's own.
- * @SuppressWarnings(PHPMD.ExcessiveClassComplexity) - Controller logic is known to have high complexity.
- * @SuppressWarnings(PHPMD.CouplingBetweenObjects) - The controller interacts with several services, resulting in high
- * coupling.
+ *
+ * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
  */
 class SelfAssertedTokensController extends Controller
 {
+    use RecoveryTokenControllerTrait;
     /**
      * @var RecoveryTokenService
      */
@@ -77,11 +70,23 @@ class SelfAssertedTokensController extends Controller
      */
     private $smsService;
 
+    /**
+     * @var LoaResolutionService
+     */
+    private $loaResolutionService;
+
+    /**
+     * @var AuthenticationRequestFactory
+     */
+    private $authnRequestFactory;
+
     public function __construct(
         RecoveryTokenService $recoveryTokenService,
         SafeStoreService $safeStoreService,
         SecondFactorService $secondFactorService,
         SmsRecoveryTokenService $smsService,
+        LoaResolutionService $loaResolutionService,
+        AuthenticationRequestFactory $authenticationRequestFactory,
         LoggerInterface $logger
     ) {
         $this->recoveryTokenService = $recoveryTokenService;
@@ -89,13 +94,18 @@ class SelfAssertedTokensController extends Controller
         $this->secondFactorService = $secondFactorService;
         $this->smsService = $smsService;
         $this->logger = $logger;
+        $this->loaResolutionService = $loaResolutionService;
+        $this->authnRequestFactory = $authenticationRequestFactory;
     }
 
     /**
-     * Select(s) the recovery token to perform the self-asserted second factor token registration with.
+     * Self-asserted token registration: Registration entrypoint
+     *
+     * Select(s) the recovery token to perform the self-asserted second factor
+     * token registration with.
      *
      * Possible outcomes:
-     * 1. Shows a recovery token selection screen when more than one tokens are available
+     * 1. Shows a recovery token selection screen when more than one token are available
      * 2. Selects the one and only available recovery token and redirects to the recovery token authentication route
      * 3. Starts registration of a recovery token if non are in possession
      */
@@ -141,8 +151,13 @@ class SelfAssertedTokensController extends Controller
     }
 
     /**
+     * Self-asserted token registration: Authenticate recovery token
+     *
      * Identity must authenticate the recovery token in order to perform a
-     * self-asserted token registration.
+     * self-asserted token registration. But only when this action is
+     * performed while recovering a Second Factor token. During initial
+     * registration of a recovery token in the self-asserted token registration
+     * flow, authentication is not required.
      */
     public function selfAssertedTokenRegistrationRecoveryTokenAction(
         Request $request,
@@ -157,6 +172,7 @@ class SelfAssertedTokensController extends Controller
 
         switch ($token->type) {
             case "sms":
+                // Todo skip authentication when recovery token is already in possession.
                 $number = InternationalPhoneNumber::fromStringFormat($token->identifier);
                 $command = new SendRecoveryTokenSmsAuthenticationChallengeCommand();
                 $command->identifier = $number;
@@ -203,6 +219,13 @@ class SelfAssertedTokensController extends Controller
         }
     }
 
+    /**
+     * Self-asserted token registration: Authenticate a SMS recovery token
+     *
+     * The previous action (selfAssertedTokenRegistrationRecoveryTokenAction)
+     * sent the Identity a OTP via SMS. The Identity must reproduce that OTP
+     * in this action. Proving possession of the recovery token.
+     */
     public function selfAssertedTokenRecoveryTokenSmsAuthenticationAction(
         Request $request,
         string $secondFactorId,
@@ -262,6 +285,13 @@ class SelfAssertedTokensController extends Controller
         );
     }
 
+    /**
+     * Self-asserted token registration: choose recovery token type
+     *
+     * The user can select which recovery token to add. Some limitations may
+     * apply. For example, using a SMS Recovery Token for registration of an
+     * SMS Second Factor is only allowed when different phone numbers are used.
+     */
     public function newRecoveryTokenAction($secondFactorId)
     {
         $this->logger->info('Determining which recovery token are available');
@@ -285,20 +315,12 @@ class SelfAssertedTokensController extends Controller
         );
     }
 
-    public function selectTokenTypeAction()
-    {
-        $this->logger->info('Determining which recovery token are available');
-        $identity = $this->getIdentity();
-        $this->assertMayAddRecoveryToken($identity);
-
-        $availableRecoveryTokens = $this->recoveryTokenService->getRemainingTokenTypes($identity);
-
-        return $this->render(
-            '@SurfnetStepupSelfServiceSelfService/registration/self_asserted_tokens/select_recovery_token.html.twig',
-            ['availableRecoveryTokens' => $availableRecoveryTokens]
-        );
-    }
-
+    /**
+     * Self-asserted token registration: Create Recovery Token (safe-store)
+     *
+     * Shows the one-time secret and asks the Identity to store the
+     * password in a safe location.
+     */
     public function registerCreateRecoveryTokenSafeStoreAction(Request $request, $secondFactorId)
     {
         $identity = $this->getIdentity();
@@ -333,6 +355,12 @@ class SelfAssertedTokensController extends Controller
         );
     }
 
+    /**
+     * Self-asserted token registration: Create the SMS recovery token
+     * Step 1: Send an OTP to phone of Identity
+     *
+     * Note: Shares logic with the recovery token SMS send challenge action
+     */
     public function registerRecoveryTokenSmsAction(Request $request, string $secondFactorId)
     {
         return $this->handleSmsChallenge(
@@ -343,6 +371,12 @@ class SelfAssertedTokensController extends Controller
         );
     }
 
+    /**
+     * Self-asserted token registration: Create the SMS recovery token
+     * Step 2: Process proof of phone possession of Identity
+     *
+     * Note: Shares logic with the recovery token SMS send challenge action
+     */
     public function registerRecoveryTokenSmsProofOfPossessionAction(Request $request, string $secondFactorId)
     {
         return $this->handleSmsProofOfPossession(
@@ -350,266 +384,6 @@ class SelfAssertedTokensController extends Controller
             '@SurfnetStepupSelfServiceSelfService/registration/self_asserted_tokens/registration_sms_prove_possession.html.twig',
             'ss_second_factor_self_asserted_tokens',
             $secondFactorId
-        );
-    }
-
-    public function createSafeStoreAction(Request $request)
-    {
-        $identity = $this->getIdentity();
-        $this->assertNoRecoveryTokenOfType('safe-store', $identity);
-        $secret = $this->safeStoreService->produceSecret();
-        $command = new PromiseSafeStorePossessionCommand();
-
-        $form = $this->createForm(PromiseSafeStorePossessionType::class, $command)->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $command->secret = $secret;
-            $command->identity = $identity;
-
-            $executionResult = $this->safeStoreService->promisePossession($command);
-            if (!$executionResult->getErrors()) {
-                return $this->redirect(
-                    $this->generateUrl('ss_second_factor_list')
-                );
-            }
-            $this->addFlash('error', 'ss.form.recovery_token.error.error_message');
-        }
-
-        return $this->render(
-            '@SurfnetStepupSelfServiceSelfService/registration/self_asserted_tokens/create_safe_store.html.twig',
-            [
-                'form' => $form->createView(),
-                'secret' => $secret,
-            ]
-        );
-    }
-
-    public function createSmsAction(Request $request)
-    {
-        return $this->handleSmsChallenge(
-            $request,
-            '@SurfnetStepupSelfServiceSelfService/registration/self_asserted_tokens/create_sms.html.twig',
-            'ss_recovery_token_prove_sms_possession'
-        );
-    }
-
-    public function proveSmsPossessionAction(Request $request)
-    {
-        return $this->handleSmsProofOfPossession(
-            $request,
-            '@SurfnetStepupSelfServiceSelfService/registration/self_asserted_tokens/sms_prove_possession.html.twig',
-            'ss_second_factor_list'
-        );
-    }
-
-    public function deleteAction(string $recoveryTokenId)
-    {
-        try {
-            $recoveryToken = $this->recoveryTokenService->getRecoveryToken($recoveryTokenId);
-            $command = new RevokeRecoveryTokenCommand();
-            $command->identity = $this->getIdentity();
-            $command->recoveryToken = $recoveryToken;
-            $executionResult = $this->safeStoreService->revokeRecoveryToken($command);
-            if (!empty($executionResult->getErrors())) {
-                $this->addFlash('error', 'ss.form.recovery_token.delete.success');
-                foreach ($executionResult->getErrors() as $error) {
-                    $this->logger->error(sprintf('Recovery Token revocation failed with message: "%s"', $error));
-                }
-                return;
-            }
-        } catch (NotFoundException $e) {
-            throw new LogicException('Identity %s tried to remove an unpossessed recovery token');
-        }
-        $this->addFlash('success', 'ss.form.recovery_token.delete.success');
-        return $this->redirect($this->generateUrl('ss_second_factor_list'));
-    }
-
-    private function assertSecondFactorInPossession(string $secondFactorId, Identity $identity)
-    {
-        $identityOwnsSecondFactor = $this->secondFactorService->identityHasSecondFactorOfStateWithId(
-            $identity->id,
-            'verified',
-            $secondFactorId
-        );
-
-        if (!$identityOwnsSecondFactor) {
-            throw new LogicException(
-                sprintf(
-                    'Identity "%s" tried to register recovery token during registration ' .
-                    'of second factor token "%s", but does not own that second factor',
-                    $identity->id,
-                    $secondFactorId
-                )
-            );
-        }
-    }
-
-    private function assertRecoveryTokenInPossession(string $recoveryTokenId, Identity $identity)
-    {
-        $recoveryTokens = $this->recoveryTokenService->getRecoveryTokensForIdentity($identity);
-        $found = false;
-        foreach ($recoveryTokens as $recoveryToken) {
-            if ($recoveryToken->recoveryTokenId === $recoveryTokenId) {
-                $found = true;
-            }
-        }
-        if (!$found) {
-            throw new LogicException(
-                sprintf(
-                    'Identity "%s" tried to perform a self-asserted token registration with a ' .
-                    'recovery token ("%s)", but does not own that recovery token',
-                    $identity->id,
-                    $recoveryTokenId
-                )
-            );
-        }
-    }
-
-    private function assertNoRecoveryTokens(Identity $identity)
-    {
-        if ($this->recoveryTokenService->hasRecoveryToken($identity)) {
-            throw new LogicException(
-                sprintf(
-                    'Identity "%s" tried to register a recovery token, but one was already in possession. ' .
-                    'This is not allowed during self-asserted token registration.',
-                    $identity->id
-                )
-            );
-        }
-    }
-
-    private function assertNoRecoveryTokenOfType(string $type, Identity $identity)
-    {
-        $tokens = $this->recoveryTokenService->getRecoveryTokensForIdentity($identity);
-        if (array_key_exists($type, $tokens)) {
-            throw new LogicException(
-                sprintf(
-                    'Identity "%s" tried to register a recovery token, but one was already in possession. ' .
-                    'This is not allowed during token registration.',
-                    $identity->id
-                )
-            );
-        }
-    }
-
-    private function assertMayAddRecoveryToken(Identity $identity)
-    {
-        $availableTypes = $this->recoveryTokenService->getRemainingTokenTypes($identity);
-        if (count($availableTypes) === 0) {
-            throw new LogicException(
-                sprintf(
-                    'Identity %s tried to register a token type, but all available token types have ' .
-                    'already been registered',
-                    $identity
-                )
-            );
-        }
-    }
-
-    /**
-     * Shared Send SMS challenge form handler
-     * - One is used during SMS recovery token registration within the vetting flow
-     * - The other is actioned from the recovery token overview on the '/overview' page
-     *
-     * Note fourth param: '$secondFactorId' is optional parameter, only used in the vetting flow scenario
-     */
-    private function handleSmsChallenge(
-        Request $request,
-        string $templateName,
-        string $exitRoute,
-        ?string $secondFactorId = null
-    ): Response {
-        $identity = $this->getIdentity();
-        $this->assertNoRecoveryTokenOfType('sms', $identity);
-        $command = new SendRecoveryTokenSmsChallengeCommand();
-        $form = $this->createForm(SendSmsChallengeType::class, $command)->handleRequest($request);
-        /** @var SmsSecondFactorService $service */
-        $otpRequestsRemaining = $this->smsService
-            ->getOtpRequestsRemainingCount(SmsSecondFactorServiceInterface::REGISTRATION_SECOND_FACTOR_ID);
-        $maximumOtpRequests = $this->smsService->getMaximumOtpRequestsCount();
-
-        $viewVariables = [
-            'otpRequestsRemaining' => $otpRequestsRemaining,
-            'maximumOtpRequests' => $maximumOtpRequests
-        ];
-
-        if (isset($secondFactorId)) {
-            $viewVariables['secondFactorId'] = $secondFactorId;
-        }
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $command->identity = $identity->id;
-            $command->institution = $identity->institution;
-
-            if ($otpRequestsRemaining === 0) {
-                $this->addFlash('error', 'ss.prove_phone_possession.challenge_request_limit_reached');
-                return array_merge(['form' => $form->createView()], $viewVariables);
-            }
-
-            if ($this->smsService->sendChallenge($command)) {
-                $urlParameter = [];
-                if (isset($secondFactorId)) {
-                    $urlParameter = ['secondFactorId' => $secondFactorId];
-                }
-                return $this->redirect($this->generateUrl($exitRoute, $urlParameter));
-            }
-            $this->addFlash('error', 'ss.form.recovery_token.error.error_message');
-        }
-        return $this->render(
-            $templateName,
-            array_merge(
-                [
-                    'form' => $form->createView(),
-                ],
-                $viewVariables
-            )
-        );
-    }
-
-    private function handleSmsProofOfPossession(
-        Request $request,
-        string $templateName,
-        string $exitRoute,
-        ?string $secondFactorId = null
-    ) {
-        if (!$this->smsService->hasSmsVerificationState(SmsRecoveryTokenService::REGISTRATION_RECOVERY_TOKEN_ID)) {
-            $this->get('session')->getFlashBag()->add('notice', 'ss.registration.sms.alert.no_verification_state');
-            return $this->redirectToRoute('ss_recovery_token_sms');
-        }
-        $identity = $this->getIdentity();
-        $this->assertNoRecoveryTokenOfType('sms', $identity);
-
-        $command = new VerifySmsRecoveryTokenChallengeCommand();
-        $command->identity = $identity->id;
-        $command->resendRoute = 'ss_recovery_token_sms';
-
-        $form = $this->createForm(VerifySmsChallengeType::class, $command)->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $result = $this->smsService->provePossession($command);
-            if ($result->isSuccessful()) {
-                $this->smsService->clearSmsVerificationState(SmsRecoveryTokenService::REGISTRATION_RECOVERY_TOKEN_ID);
-                $urlParameter = [];
-                if (isset($secondFactorId)) {
-                    $urlParameter = ['secondFactorId' => $secondFactorId];
-                }
-                return $this->redirect($this->generateUrl($exitRoute, $urlParameter));
-            } elseif ($result->wasIncorrectChallengeResponseGiven()) {
-                $this->addFlash('error', 'ss.prove_phone_possession.incorrect_challenge_response');
-            } elseif ($result->hasChallengeExpired()) {
-                $this->addFlash('error', 'ss.prove_phone_possession.challenge_expired');
-            } elseif ($result->wereTooManyAttemptsMade()) {
-                $this->addFlash('error', 'ss.prove_phone_possession.too_many_attempts');
-            } else {
-                $this->addFlash('error', 'ss.prove_phone_possession.proof_of_possession_failed');
-            }
-        }
-
-        return $this->render(
-            $templateName,
-            [
-                'form' => $form->createView(),
-            ]
         );
     }
